third_party/blink/web_tests/external/wpt/digital-credentials/identity-create.tentative.https-expected.txt

This is a testharness.js-based test.
[FAIL] navigator.identity.create() frame just returns null.
  promise_test: Unhandled rejection with value: object "NotSupportedError: Failed to execute 'create' on 'CredentialsContainer': Create operation not supported for this credential type."
[FAIL] navigator.identity.create() same-origin iframe just returns null.
  promise_test: Unhandled rejection with value: object "NotSupportedError: Failed to execute 'create' on 'CredentialsContainer': Create operation not supported for this credential type."
[FAIL] navigator.identity.create() cross-origin iframe results in null.
  assert_equals: expected null but got object "[object Object]"
[FAIL] navigator.identity.create() ignores mediations.
  promise_test: Unhandled rejection with value: object "NotSupportedError: Failed to execute 'create' on 'CredentialsContainer': Create operation not supported for this credential type."
[FAIL] navigator.identity.create() ignores mediations in same-origin iframe.
  promise_test: Unhandled rejection with value: object "NotSupportedError: Failed to execute 'create' on 'CredentialsContainer': Create operation not supported for this credential type."
[FAIL] navigator.identity.create() ignores mediations in cross-origin iframe.
  assert_equals: expected null but got object "[object Object]"
[FAIL] navigator.identity.create() rejects if called with an aborted signal.
  promise_rejects_dom: function "function() { throw e }" threw object "NotSupportedError: Failed to execute 'create' on 'CredentialsContainer': Create operation not supported for this credential type." that is not a DOMException AbortError: property "code" is equal to 9, expected 20
[FAIL] navigator.identity.create() rejects if called with an aborted signal in same-origin iframe.
  promise_rejects_dom: function "function() { throw e }" threw object "NotSupportedError: Failed to execute 'create' on 'CredentialsContainer': Create operation not supported for this credential type." that is not a DOMException AbortError: property "code" is equal to 9, expected 20
[FAIL] navigator.identity.create() rejects if called with an aborted signal in cross-origin iframe.
  assert_equals: expected "AbortError" but got "NotSupportedError"
[FAIL] Adding mediations together with abort signal respects the abort signal.
  promise_rejects_dom: function "function() { throw e }" threw object "NotSupportedError: Failed to execute 'create' on 'CredentialsContainer': Create operation not supported for this credential type." that is not a DOMException AbortError: property "code" is equal to 9, expected 20
Harness: the test ran to completion.
